settings: minor code cleanup
 - separate internal and values from the public APIs.

 - rename `WritableSetting` to `NonMaskedSetting` ("writable" will
   become confusing when we will have "tenant read-only" and "tenant
   writable" settings).

 - rename `extendedSetting` to `internalSetting`.

Release note: None
